Intractive legend toggle UI update
