+++ index.html
@@ -1,10 +1,3 @@
-<!-- 
-  generatePageContent.tsx looks for the following strings in this file:
-  - "TITLE PLACEHOLDER"
-  - "<style></style>"
-  - "CONTENT PLACEHOLDER"
-  If you change any of these, make sure to edit generatePageContent.tsx as well. -->
-
 <html>
   <head>
     <meta charset="utf-8" />
