+++ src/impact-dashboard/ChartArea.tsx
@@ -1,15 +1,9 @@
-// @ts-nocheck
-import { zip } from "d3-array";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import styled from "styled-components";
 
 import Colors from "../design-system/Colors";
-import { calculateCurves, CurveData } from "../infection-model";
-import { getAllValues, getColView } from "../infection-model/matrixUtils";
-import { seirIndex } from "../infection-model/seir";
 import CurveChart from "./CurveChartContainer";
 import CurveChartLegend from "./CurveChartLegend";
-import { useEpidemicModelState } from "./EpidemicModelContext";
 
 export type MarkColors = {
   exposed: string;
@@ -19,7 +13,7 @@ export type MarkColors = {
   hospitalBeds: string;
 };
 
-interface GroupStatus {
+interface GroupStatusProps {
   [propName: string]: boolean;
 }
 
@@ -36,25 +30,7 @@ const LegendContainer = styled.div`
   flex: 0 0 auto;
 `;
 
-function combinePopulations(data: CurveData, columnIndex: number) {
-  return zip(
-    getAllValues(getColView(data.incarcerated, columnIndex)),
-    getAllValues(getColView(data.staff, columnIndex)),
-  ).map(([incarcerated, staff]) => incarcerated + staff);
-}
-
-const ChartArea: React.FC<> = () => {
-  const modelData = useEpidemicModelState();
-
-  const projectionData = calculateCurves(modelData);
-
-  const curveData = {
-    exposed: combinePopulations(projectionData, seirIndex.exposed),
-    fatalities: combinePopulations(projectionData, seirIndex.fatalities),
-    hospitalized: combinePopulations(projectionData, seirIndex.hospitalized),
-    infectious: combinePopulations(projectionData, seirIndex.infectious),
-  };
-
+const ChartArea: React.FC<GroupStatusProps> = () => {
   const markColors = {
     exposed: Colors.green,
     fatalities: Colors.black,
@@ -64,33 +40,15 @@ const ChartArea: React.FC<> = () => {
   };
 
   const [groupStatus, setGroupStatus] = useState({
-    exposed: false,
-    fatalities: false,
-    hospitalized: false,
-    infectious: false,
+    exposed: true,
+    fatalities: true,
+    hospitalized: true,
+    infectious: true,
   });
-
-  const [displayData, setDisplayData] = useState({});
 
   const toggleGroup = (groupName: keyof typeof groupStatus) => {
     setGroupStatus({ ...groupStatus, [groupName]: !groupStatus[groupName] });
   };
-
-  const updateDisplayData = () => {
-    let filteredGroupStatus = Object.keys(groupStatus).filter(
-      (groupName) => groupStatus[groupName],
-    );
-    setDisplayData(
-      filteredGroupStatus.reduce(
-        (data, key) => Object.assign(data, { [key]: curveData[key] }),
-        {},
-      ),
-    );
-  };
-
-  useEffect(() => {
-    updateDisplayData();
-  }, [groupStatus]);
 
   return (
     <Container>
@@ -103,7 +61,7 @@ const ChartArea: React.FC<> = () => {
           />
         </LegendContainer>
       </LegendAndActions>
-      <CurveChart markColors={markColors} curveData={displayData} />
+      <CurveChart markColors={markColors} groupStatus={groupStatus} />
     </Container>
   );
 };
