+++ src/impact-dashboard/CurveChart.tsx
@@ -137,7 +137,6 @@ const CurveChart: React.FC<CurveChartProps> = ({
   curveData,
   hospitalBeds,
   markColors,
-  // toggled,
 }) => {
   const frameProps = {
     lines: Object.entries(curveData).map(([bucket, values]) => ({
